+++ exec/Kathu/Editor/Panels.hs
@@ -4,22 +4,108 @@
 
 module Kathu.Editor.Panels where
 
-import           Control.Monad              (forM_, void)
+import qualified Apecs
+import           Control.Lens               hiding (set)
+import           Control.Monad              (forM_, void, when)
 import           Data.GI.Base
 import           Data.IORef
+import qualified Data.Map                   as Map
 import           Data.Text                  (Text)
+import qualified Data.Text                  as T
 import           Data.Vector                (Vector)
 import qualified Data.Vector                as Vec
 import qualified GI.Gtk                     as Gtk
+import qualified GI.GdkPixbuf               as Gdk
 
+import           Kathu.App.Data.Library     (kathuStore, tiles)
+import           Kathu.App.Data.KathuStore  (countingIDs)
+import           Kathu.App.Graphics.Image   (ImageID(..))
 import           Kathu.App.Tools.EventQueue
 import           Kathu.App.Tools.ToolMode
 import           Kathu.Editor.GtkUtil
 import           Kathu.Editor.Resources
 import           Kathu.Editor.Types
+import           Kathu.Graphics.Drawable    (getRenderGraphicsVector)
+import           Kathu.Parsing.Counting
+import           Kathu.Util.Collection      (fromJustElseError)
 import           Kathu.Util.Types
 import           Kathu.World.WorldSpace
-import           Kathu.World.Tile           (emptyTile)
+import           Kathu.World.Tile           (emptyTile, emptyTileID, tileID, tileName, tileRender)
+
+mkTileIcon :: Text -> IO Gtk.Image
+mkTileIcon path = do
+    fullImg <- Gdk.pixbufNewFromFile (T.unpack path)
+    width   <- Gdk.pixbufGetWidth  fullImg
+    height  <- Gdk.pixbufGetHeight fullImg
+
+    let fromJustElseMemError = fromJustElseError "No more memory could be allocated for Tile Pixbuf"
+
+    croppedPixBuf <-
+        if   width < 16 || height < 16
+        then error $ "Tile icon for path " ++ show path ++ " is too small (less than 16x16)"
+        else if width == 16 && height == 16
+            then pure fullImg
+            else do
+                -- take bottom left 16x16
+                pixbuf <- fromJustElseMemError <$> Gdk.pixbufNew Gdk.ColorspaceRgb True 8 16 16
+                Gdk.pixbufCopyArea fullImg 0 (height - 16) 16 16 pixbuf 0 0
+                pure pixbuf
+    finalPixbuf <- fromJustElseMemError <$> Gdk.pixbufScaleSimple croppedPixBuf 32 32 Gdk.InterpTypeNearest
+    Gtk.imageNewFromPixbuf (Just finalPixbuf)
+
+mkTileSelectorPanel :: EditorState -> IO Gtk.Widget
+mkTileSelectorPanel EditorState{eventQueue = queue} = do
+    flowbox <- new Gtk.FlowBox [#activateOnSingleClick := True, #maxChildrenPerLine := 10]
+
+    library <- runWithEntityWorld queue $ Apecs.get Apecs.global
+    let -- remove emptyTile, since it has no graphics and will error if we try to use them
+        libTiles          = library^.tiles.to (filter (\t -> t^.tileID /= emptyTileID) . Map.elems)
+        libTilesByImageID = Map.fromList . map (\t -> (getImageID t, t)) $ libTiles
+        
+        getImageID t      = t^.tileRender.to (Vec.head . getRenderGraphicsVector)
+        isTileImage imgID  = Map.member imgID libTilesByImageID
+
+        tileImageCounting = library^.kathuStore.countingIDs.to ((Map.! "ImageID") . unCounting)
+
+        tileImagePaths    = Map.fromList $ Map.foldlWithKey' appendIfTileImage [] tileImageCounting
+        appendIfTileImage acc path idx
+            | isTileImage imgID = (imgID, path):acc
+            | otherwise         = acc
+            where imgID = ImageID $ fromIntegral idx
+
+    emptyTileIcon <- mkTileIcon "./assets/editor/empty-tile-icon.png"
+    tileIcons     <- mapM mkTileIcon tileImagePaths
+
+    let tileIconPairs = (emptyTile, emptyTileIcon) : Map.foldlWithKey' addPair [] libTilesByImageID
+        -- don't add tiles without any found sprites
+        addPair acc imgID tile = case Map.lookup imgID tileIcons of
+            Just icon -> (tile, icon):acc
+            Nothing   -> acc
+        tilesVec      = Vec.fromList . map fst $ tileIconPairs
+
+    forM_ tileIconPairs $ \(_, icon) ->
+        Gtk.containerAdd flowbox icon
+
+    Gtk.containerForeach flowbox $ \child -> do
+        fbChild <- unsafeCastTo Gtk.FlowBoxChild child
+        idx     <- fromIntegral <$> Gtk.flowBoxChildGetIndex fbChild
+        set child [#tooltipText := (tilesVec Vec.! idx)^.tileName]
+        when (idx == 0) $
+            Gtk.flowBoxSelectChild flowbox fbChild
+
+    void $ on flowbox #childActivated $ \child -> do
+        idx  <- fromIntegral <$> Gtk.flowBoxChildGetIndex child
+        if   idx < 0 || idx >= Vec.length tilesVec
+        then putStrLn $ "Tile selector flowbox child widget was activated, but had an index out of range (was "
+                     ++ show idx
+                     ++  ", must be in (0, "
+                     ++ show (Vec.length tilesVec)
+                     ++ ")"
+        else pushAppEvent queue $ SetSelectedTile (tilesVec Vec.! idx)
+
+    frame <- new Gtk.Frame [#label := "Tiles"]
+    Gtk.containerAdd frame flowbox
+    Gtk.toWidget frame
 
 mkWorldSpaceToolbar :: EditorState -> IO Gtk.Toolbar
 mkWorldSpaceToolbar EditorState{eventQueue = queue, resources = res} = do
@@ -27,8 +113,9 @@ mkWorldSpaceToolbar EditorState{eventQueue = queue, resources = res} = do
 
     let btnConfigs :: Vector (Text, ToolMode, Gtk.Image, Gtk.Image)
         btnConfigs = Vec.fromList
-            [ ("Play Game",  NoTool, iconToolPlayGame res, iconToolPlayGameActive res)
-            , ("Draw Tiles", TilePlacer (mkTilePlacerState emptyTile), iconToolTilePlacer res, iconToolTilePlacerActive res)
+            [ ("Play Game",    NoTool, iconToolPlayGame res, iconToolPlayGameActive res)
+            , ("Draw Tiles",   TilePlacer emptyTilePlacerState, iconToolTilePlacer res, iconToolTilePlacerActive res)
+            , ("Wire Signals", SignalWirer, iconToolSignalWirer res, iconToolSignalWirerActive res)
             ]
         mkButton idx (lbl, mode, icon, activeIcon) = do
             img <- Gtk.imageNewFromPixbuf =<< Gtk.imageGetPixbuf (if idx == 0 then activeIcon else icon)
@@ -70,5 +157,8 @@ mkWorldSpacePanel es@EditorState{wsEditState = wsStateRef} = do
 
     toolbar <- mkWorldSpaceToolbar es
     Gtk.containerAdd box toolbar
+
+    tileSelector <- mkTileSelectorPanel es
+    Gtk.containerAdd box tileSelector
 
     Gtk.toWidget box