+++ exec/Kathu/Editor/Resources.hs
@@ -11,15 +11,19 @@ appIconPath :: String
 appIconPath = editorAssetPath ++ "icon-large.png"
 
 data Resources = Resources
-    { iconToolPlayGame         :: Gtk.Image
-    , iconToolPlayGameActive   :: Gtk.Image
-    , iconToolTilePlacer       :: Gtk.Image
-    , iconToolTilePlacerActive :: Gtk.Image
+    { iconToolPlayGame          :: Gtk.Image
+    , iconToolPlayGameActive    :: Gtk.Image
+    , iconToolSignalWirer       :: Gtk.Image
+    , iconToolSignalWirerActive :: Gtk.Image
+    , iconToolTilePlacer        :: Gtk.Image
+    , iconToolTilePlacerActive  :: Gtk.Image
     }
 
 loadResources :: IO Resources
 loadResources = Resources
     <$> Gtk.imageNewFromFile (editorAssetPath ++ "tool-play-game.png")
     <*> Gtk.imageNewFromFile (editorAssetPath ++ "tool-play-game-active.png")
+    <*> Gtk.imageNewFromFile (editorAssetPath ++ "tool-signal-wirer.png")
+    <*> Gtk.imageNewFromFile (editorAssetPath ++ "tool-signal-wirer-active.png")
     <*> Gtk.imageNewFromFile (editorAssetPath ++ "tool-tile-placer.png")
     <*> Gtk.imageNewFromFile (editorAssetPath ++ "tool-tile-placer-active.png")